% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc.ppmnet}
\alias{auc.ppmnet}
\title{AUC values for a 'ppmnet' object}
\usage{
\method{auc}{ppmnet}(X, data, s = NULL, ...)
}
\arguments{
\item{X}{A fitted \code{ppmnet} object.}

\item{data}{A list of pixel images (of class \code{imlist})
containing the spatial covariates used to fit the model.}

\item{s}{Value(s) of the penalty tuning parameter at which AUC is to
be calculated. Default is the entire sequence used to fit the
regularization path.}

\item{...}{Additional arguments passed to \code{predict.ppmnet}.}
}
\description{
Computes the area under the receiver operating characteristic curve (AUC) for
a regularization path of point process models.
}
\examples{
Qp <- quadscheme(Xp)
fitp <- ppmnet(Qp, exdata, nlambda = 10)
auc(fitp, exdata)

}
