% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmnet.R
\name{deviance.ppmnet}
\alias{deviance.ppmnet}
\title{Extract the deviance from a 'ppmnet' object}
\usage{
\method{deviance}{ppmnet}(object, ...)
}
\arguments{
\item{object}{A fitted \code{ppmnet} object.}

\item{...}{Ignored.}
}
\description{
Extract the deviance from a 'ppmnet' object
}
