% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exdata}
\alias{exdata}
\title{Example spatial covariate data}
\format{A list of pixel images containing simulated spatial covariates.}
\usage{
exdata
}
\description{
Simulated example spatial covariate data.
}
\keyword{datasets}
