% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.ppmnet}
\alias{fitted.ppmnet}
\title{Fitted conditional intensity from a 'ppmnet' object}
\usage{
\method{fitted}{ppmnet}(object, s = NULL, drop = FALSE, ...)
}
\arguments{
\item{object}{A fitted \code{ppmnet} object.}

\item{s}{Value(s) of the penalty tuning parameter at which predictions are to
be made. Default is the entire sequence used to fit the regularization
path.}

\item{drop}{Logical value. If \code{TRUE}, quadrature points that were not
used to fit the model are deleted.}

\item{...}{Ignored}
}
\value{
A matrix of fitted values, or, if \code{s} is of length 1, a vector
        of fitted values.
}
\description{
Computes the fitted conditional intensity for a regularized spatial point
process model at the quadrature points used to fit the model.
}
\examples{
Qp <- quadscheme(Xp)
fitp <- ppmnet(Qp, exdata)
fitted(fitp)

}
