% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.ppmnet}
\alias{logLik.ppmnet}
\title{Log (composite) likelihood for a 'ppmnet' object}
\usage{
\method{logLik}{ppmnet}(object, s = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{ppmnet} object.}

\item{s}{Value(s) of the penalty tuning parameter at which log (composite)
likelihood is to be computed. Default is the entire sequence used to
fit the regularization path.}

\item{...}{Ignored.}
}
\value{
A numeric vector.
}
\description{
Extracts the log likelihood for a regularized Poisson point process model, or
the log composite likelihood for a regularized Gibbs point process model.
}
\examples{
Qp <- quadscheme(Xp)
fitp <- ppmnet(Qp, exdata)
logLik(fitp)

}
