% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmnet.R
\name{plot.ppmnet}
\alias{plot.ppmnet}
\title{Plot the coefficients from a 'ppmnet' object}
\usage{
\method{plot}{ppmnet}(x, ...)
}
\arguments{
\item{x}{A fitted \code{ppmnet} object.}

\item{...}{Additional arugments passed to \code{plot.glmnet}.}
}
\description{
Plot the coefficients from a 'ppmnet' object
}
