% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmnet.R
\name{ppmnet}
\alias{ppmnet}
\title{Fit a spatial point process model with lasso or elastic net regularization}
\usage{
ppmnet(Q, data, interaction = NULL, correction = "border",
  rbord = NULL, method = "mpl", ...)
}
\arguments{
\item{Q}{A quadrature scheme (of class \code{quad}) containing a
point pattern.}

\item{data}{A list of pixel images (of class \code{imlist})
representing spatial covariates.}

\item{interaction}{An object (of class \code{interact}) describing the
point process interaction structure, or \code{NULL} indicating that a
Poisson point process model is to be fit.}

\item{correction}{Edge correction to be used. Either "\code{border}" for
border correction, or "\code{none}" for no correction.}

\item{rbord}{If \code{correction = "border"}, the distance by which the
observation window is eroded.}

\item{method}{Method to be used to fit the model. Either
"\code{mpl}" for penalized maximum pseudolikelihood, or "\code{logi}"
for penalized maximum logistic composite likelihood.}

\item{...}{Additional arguments passed to \code{glmnet} to control the
fitting procedure.}
}
\value{
An object of class "\code{ppmnet}".
}
\description{
Fits spatial point process models via penalized composite likelihood. The
regularization path is computed for the lasso or elastic net penalty along a
sequence of tuning parameter values. Support for Poisson point process models
and for Gibbs point process models.
}
\examples{
# Poisson model fit via penalized maximum likelihood
Qp <- spatstat::quadscheme(Xp)
fitp <- ppmnet(Qp, exdata)

# Strauss model fit via penalized maximum pseudolikelihood
Qs <- spatstat::quadscheme(Xs)
fits <- ppmnet(Qs, exdata, interaction = spatstat::Strauss(5), nlambda = 20)

# Geyer saturation model fit via penalized logistic composite likelihood
Qg <- spatstat::quadscheme.logi(Xg)
fitg <- ppmnet(Qg, exdata, interaction = spatstat::Geyer(5, 1),
               method = "logi", nlambda = 20)

}
