% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmnet.R
\name{print.ppmnet}
\alias{print.ppmnet}
\title{Print a 'ppmnet' object}
\usage{
\method{print}{ppmnet}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A fitted \code{ppmnet} object.}

\item{digits}{The number of significant digits used in the printout.}

\item{...}{Ignored.}
}
\description{
Print a 'ppmnet' object
}
