% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.ppmnet}
\alias{select.ppmnet}
\alias{select}
\title{Select a tuning parameter value from a 'ppmnet' object}
\usage{
\method{select}{ppmnet}(object, criterion = c("AIC", "BIC", "ERIC"), ...)
}
\arguments{
\item{object}{A fitted \code{ppmnet} object.}

\item{criterion}{The criterion by which to select the tuning parameter. One
of \code{"AIC"}, \code{"BIC"}, or \code{"ERIC"}.}

\item{...}{Ignored}
}
\value{
A list containing \code{lambda}, the selected tuning parameter;
        \code{coefs}, the coefficients of the selected model; \code{edf}, the
        effective degrees of freedom for the regularization path; and
        \code{criterion}; the computed criterion values for the
        regularization path.
}
\description{
Selects a point from along the fitted regularization path of a 'ppmnet'
object on the basis of a (composite) information criterion.
}
\examples{
Qp <- quadscheme(Xp)
fitp <- ppmnet(Qp, exdata)
select(fitp, "BIC")

Qg <- quadscheme.logi(Xg)
fitg <- ppmnet(Qg, exdata, Geyer(5, 1), method = "logi", nlambda = 20)
select(fitg, "BIC")

}
