% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xg}
\alias{Xg}
\title{Simulated inhomogeneous Geyer point pattern}
\format{An object of class \code{ppp}.}
\usage{
Xg
}
\description{
Simulated inhomogeneous Geyer point pattern.
}
\keyword{datasets}
