% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xp}
\alias{Xp}
\title{Simulated inhomogeneous Poisson point pattern}
\format{An object of class \code{ppp}.}
\usage{
Xp
}
\description{
Simulated inhomogeneous Poisson point pattern.
}
\keyword{datasets}
