% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xs}
\alias{Xs}
\title{Simulated inhomogeneous Strauss point pattern}
\format{An object of class \code{ppp}.}
\usage{
Xs
}
\description{
Simulated inhomogeneous Strauss point pattern.
}
\keyword{datasets}
