% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.ppmnet}
\alias{select.ppmnet}
\alias{select}
\title{Select a tuning parameter value from a 'ppmnet' object}
\usage{
\method{select}{ppmnet}(object, criterion = c("AIC", "BIC", "ERIC"), ...)
}
\arguments{
\item{object}{A fitted \code{ppmnet} object.}

\item{criterion}{The criterion by which to select the tuning parameter. One
of \code{"AIC"}, \code{"BIC"}, or \code{"ERIC"}.}

\item{...}{Ignored}
}
\description{
Selects a point from along the fitted regularization path of a 'ppmnet'
object on the basis of a (composite) information criterion.
}
